styling updates, transitioning styles from style.dart to themeDart
