+++ lib/ui/appTheme.dart
@@ -1,31 +1,32 @@
 import 'package:flutter/material.dart';
 
 
-var appColors = LightThemeColors();
+var lightAppColors = LightThemeColors();
+var darkAppColors = DarkThemeColors();
 
 
 class AppTheme {
   AppTheme._();
 
   static final ThemeData lightTheme = ThemeData(
-    scaffoldBackgroundColor: Color(0xffFFFFFF),
+    scaffoldBackgroundColor: Color(0xffF3F3F3),
     appBarTheme: AppBarTheme(
-      color: Color(0xffFFFFFF),
+      color: Color(0xffF3F3F3),
       iconTheme: IconThemeData(
-        color: appColors.text,
+        color: lightAppColors.text,
       ),
     ),
     backgroundColor: Color(0xffFFFFFF),
     colorScheme: ColorScheme.light(
-      primary: Color(0xff4D5CD0),
+      primary: Color(0xff5468f7),
       onPrimary: Colors.white,
-      primaryVariant:  Color(0xffBFD9FF),
+      primaryVariant:  Color(0xff354355),
       secondary: Colors.red,
-      surface: Color(0xffF3F3F3)
+      surface: Color(0xffFFFFFF)
     ),
     cardTheme: CardTheme(
-      color: Color(0xFFF3F3F3),
-      elevation: 2.0,
+      color: Color(0xFFFFFFFF),
+      elevation: 3.0,
       margin: EdgeInsets.all(20.0)
     ),
     iconTheme: IconThemeData(
@@ -33,64 +34,96 @@ class AppTheme {
     ),
     textTheme: TextTheme(
       headline4: TextStyle(
-        color: appColors.text,
+        color: lightAppColors.text,
         fontSize: 28.0,
       ),
       headline5: TextStyle(
-        color: appColors.text,
+        color: lightAppColors.text,
         fontWeight: FontWeight.bold,
         fontSize: 22.0,
       ),
       headline6: TextStyle(
-        color: appColors.text,
+        color: lightAppColors.text,
         fontWeight: FontWeight.bold,
-        fontSize: 20.0,
+        fontSize: 19.0,
       ),
       subtitle2: TextStyle(
-        color: appColors.text,
+        color: lightAppColors.text,
         fontSize: 18.0,
       ),
       subtitle1: TextStyle(
-        color: appColors.text,
+        color: lightAppColors.text,
         fontSize: 18.0,
       ),
       bodyText1: TextStyle(
-        color: appColors.textSecondary,
+        color: lightAppColors.textSecondary,
+        height: 1.4,
         fontSize: 16.0,
       ),
       bodyText2: TextStyle(
-        color: appColors.text,
+        color: lightAppColors.text,
+        height: 1.4,
         fontSize: 14.0,
       ),
     ),
   );
 
   static final ThemeData darkTheme = ThemeData(
-    scaffoldBackgroundColor: Colors.black,
+    scaffoldBackgroundColor: Color(0xFF212529),
     appBarTheme: AppBarTheme(
-      color: Colors.black,
+      color: Color(0xFF212529),
       iconTheme: IconThemeData(
         color: Colors.white,
       ),
     ),
+    backgroundColor: Color(0xFF202125),
     colorScheme: ColorScheme.light(
-      primary: Colors.black,
-      onPrimary: Colors.black,
-      primaryVariant: Colors.black,
+      primary: Color(0xFFE0E6E9),
+      onPrimary: Colors.white,
+      primaryVariant:  Color(0xff354355),
       secondary: Colors.red,
+      surface: Color(0xFF2a2e32)
     ),
-    cardTheme: CardTheme(color: Colors.black, elevation: 2.0),
+    cardTheme: CardTheme(
+      color: Color(0xFF2a2e32),
+      elevation: 3.0,
+      margin: EdgeInsets.all(20.0)
+    ),
     iconTheme: IconThemeData(
       color: Colors.white54,
     ),
     textTheme: TextTheme(
+      headline4: TextStyle(
+        color: darkAppColors.text,
+        fontSize: 28.0,
+      ),
+      headline5: TextStyle(
+        color: darkAppColors.text,
+        fontWeight: FontWeight.bold,
+        fontSize: 22.0,
+      ),
       headline6: TextStyle(
         color: Colors.white,
-        fontSize: 20.0,
+        fontWeight: FontWeight.bold,
+        fontSize: 19.0,
       ),
       subtitle2: TextStyle(
-        color: Colors.white70,
+        color: darkAppColors.text,
         fontSize: 18.0,
+      ),
+      subtitle1: TextStyle(
+        color: darkAppColors.text,
+        fontSize: 18.0,
+      ),
+      bodyText1: TextStyle(
+        color: darkAppColors.textSecondary,
+        height: 1.4,
+        fontSize: 16.0,
+      ),
+      bodyText2: TextStyle(
+        color: darkAppColors.text,
+        height: 1.4,
+        fontSize: 14.0,
       ),
     ),
   );
@@ -106,5 +139,16 @@ class LightThemeColors {
   LightThemeColors() {
       text = Color(0xff797979);
       textSecondary = Color(0xffA4A4A4);
+  }
+}
+
+class DarkThemeColors {
+
+  Color text;
+  Color textSecondary;
+
+  DarkThemeColors() {
+      text = Colors.white;
+      textSecondary = Color(0xFF6f7275);
   }
 }
