+++ lib/ui/views/settings.dart
@@ -13,11 +13,6 @@ class SettingsPage extends StatefulWidget {
 class _SettingsPageState extends State<SettingsPage> {
   @override
   Widget build(BuildContext context) {
-    if (darkMode) {
-      appColors = AppColors(0);
-    } else {
-      appColors = AppColors(1);
-    }
     return Scaffold(
       body: Center(
         child: Container(
