+++ lib/ui/widgets/custom_widgets.dart
@@ -1,219 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:voting_app/core/models/bill.dart';
 import 'package:voting_app/ui/styles.dart';
-import 'package:pie_chart/pie_chart.dart';
 import 'dart:async';
-
-class HouseIconsWidget extends StatelessWidget {
-
-  final Bill bill;
-  final Color senateColor = appColors.senate;
-  final Color houseColor = appColors.house;
-  final Color noFillColor = appColors.greyedOut;
-  final double size;
-
-  /// shows the progression of a bill as coloured icons
-  ///
-  /// colours represent the house and senate
-  /// usage:
-  ///
-  /// `child: HouseIconsWidget(issuesMap: issuesMap,size: 20,),`
-  HouseIconsWidget({
-    Key key,
-    @required this.size,
-    @required this.bill
-  }) : super(key: key);
-
-  // gets correct colour for Intro House
-  hiChooser(theBill) {
-    if (theBill.chamber == "House") {
-      if (theBill.introHouse == "") {
-        return noFillColor;
-      } else {
-        return houseColor;
-      }
-    } else {
-      if (theBill.introSenate == "") {
-        return noFillColor;
-      } else {
-        return senateColor;
-      }
-    }
-  }
-
-  // gets correct colour for Passed House
-  hpChooser(theBill) {
-    if (theBill.chamber == "House") {
-      if (theBill.passedHouse == "") {
-        return noFillColor;
-      } else {
-        return houseColor;
-      }
-    } else {
-      if (theBill.passedSenate == "") {
-        return noFillColor;
-      } else {
-        return senateColor;
-      }
-    }
-  }
-
-  // gets correct colour for Intro Senate
-  siChooser(theBill) {
-    if (theBill.chamber == "House") {
-      if (theBill.introHouse == "") {
-        return noFillColor;
-      } else {
-        return senateColor;
-      }
-    } else {
-      if (theBill.introHouse == "") {
-        return noFillColor;
-      } else {
-        return houseColor;
-      }
-    }
-  }
-
-  // gets correct colour for Passed Senate
-  spChooser(theBill) {
-    if (theBill.chamber == "House") {
-      if (theBill.passedSenate == "") {
-        return noFillColor;
-      } else {
-        return senateColor;
-      }
-    } else {
-      if (theBill.passedHouse == "") {
-        return noFillColor;
-      } else {
-        return houseColor;
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // icons
-    return Container(
-      margin: EdgeInsets.all(appSizes.standardPadding),
-      height: this.size,
-      width: this.size * 7,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.start,
-        children: <Widget>[
-          Icon(
-            Icons.account_balance,
-            color: hiChooser(bill),
-            size: this.size,
-          ),
-          Icon(
-            Icons.label_important,
-            color: appColors.greyedOut,
-            size: this.size,
-          ),
-          Icon(
-            Icons.check_circle,
-            color: hpChooser(bill),
-            size: this.size,
-          ),
-          Icon(
-            Icons.label_important,
-            color: appColors.greyedOut,
-            size: this.size,
-          ),
-          Icon(
-            Icons.account_balance,
-            color: siChooser(bill),
-            size: this.size,
-          ),
-          Icon(
-            Icons.label_important,
-            color: appColors.greyedOut,
-            size: this.size,
-          ),
-          Icon(
-            Icons.check_circle,
-            color: spChooser(bill),
-            size: this.size,
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class VotingStatusWidget extends StatelessWidget {
-  final Bill bill;
-  final bool voted;
-  final double size;
-
-  ///Widget showing the voting status
-  ///
-  /// displays as either: voting open, voting closed or voted
-  ///
-  /// Usage:
-  ///
-  /// `child:  VotingStatusWidget(issuesMap: issuesMap,voted: true,size: 20),`
-
-  VotingStatusWidget({
-    Key key,
-    @required this.bill,
-    @required this.voted,
-    @required this.size,
-  }) : super(key: key);
-
-  /// To get the status message format
-  ///
-  /// Returns a list: [colour, message, icon]
-  statusMessage() {
-    String s = "Closed";
-    Color c = appColors.voteClosed;
-    var i = Icons.adjust;
-    if (voted) {
-      s = "Voted";
-      c = appColors.voted;
-      i = Icons.check_circle_outline;
-    } else {
-      if (bill.chamber == "House") {
-        if (bill.passedSenate == "") {
-          s = "Open";
-          c = appColors.voteOpen;
-          i = Icons.add_circle_outline;
-        }
-      } else {
-        if (bill.passedHouse == "") {
-          s = "Open";
-          c = appColors.voteOpen;
-          i = Icons.add_circle_outline;
-        }
-      }
-    }
-    // [colour, message, icon]
-    return [c, s, i];
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-        child: Column(
-      children: <Widget>[
-        Icon(
-          statusMessage()[2], // icon
-          color: statusMessage()[0], //color
-          size: this.size, // size
-        ),
-        Text(
-          statusMessage()[1], // message
-          style: TextStyle(
-              fontSize: this.size * 4 / 10,
-              fontWeight: FontWeight.bold,
-              color: statusMessage()[0]),
-        ),
-      ],
-    ));
-  }
-}
 
 class CountUpWidget extends StatefulWidget {
   @override
@@ -271,47 +58,6 @@ class _CountUpWidgetState extends State<CountUpWidget> {
         ],
       ),
     ));
-  }
-}
-
-class PieWidget extends StatelessWidget {
-  final int yes;
-  final int no;
-  final double radius;
-  final bool showValues;
-
-  /// Pie graph showing yes vs no
-  ///
-  /// usage:
-  ///
-  /// `child: PieWidget(yes: 1000, no: 551, radius: 50,),`
-  PieWidget({
-    Key key,
-    @required this.yes,
-    @required this.no,
-    @required this.radius,
-    @required this.showValues
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    double yesOut = (this.yes / (this.yes + this.no)) * 100;
-    double noOut = 100 - yesOut;
-
-    Map<String, double> dataMap = new Map();
-    dataMap.putIfAbsent("No", () => noOut);
-    dataMap.putIfAbsent("Yes", () => yesOut);
-
-    return PieChart(
-      dataMap: dataMap,
-      animationDuration: Duration(milliseconds: 800),
-      chartLegendSpacing: 10.0,
-      chartRadius: this.radius,
-      showChartValues: false,
-      showChartValueLabel: false,
-      legendStyle:
-          TextStyle(fontSize: this.radius / 20 + 10, color: appColors.text),
-    );
   }
 }
