+++ lib/ui/widgets/issue_list_item.dart
@@ -2,11 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:voting_app/core/models/issue.dart';
 import 'dart:math';
 import 'package:voting_app/ui/styles.dart';
-import 'package:voting_app/ui/widgets/custom_widgets.dart';
 import 'package:voting_app/ui/views/issues/issue.dart';
+import 'package:voting_app/ui/widgets/pie_chart.dart';
 
 class IssueListItem extends StatelessWidget {
-
   final Issue issue;
   final Map billColorsDark = {
     "House": appColors.house,
@@ -29,37 +28,32 @@ class IssueListItem extends StatelessWidget {
             // Pushing a named route
             Navigator.push(
               context,
-              MaterialPageRoute(
-                  builder: (context) => IssuePage(issue: issue)),
+              MaterialPageRoute(builder: (context) => IssuePage(issue: issue)),
             );
           },
           child: Container(
-            width: 500,
+            width: appSizes.mediumWidth,
+            padding: EdgeInsets.all(appSizes.standardPadding),
             child: Column(
               mainAxisAlignment: MainAxisAlignment.spaceAround,
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: <Widget>[
                 Container(
-                  padding: EdgeInsets.all(10),
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.spaceBetween,
                     children: <Widget>[
                       Text("Open",
-                          style: TextStyle(
-                              fontSize: 13,
-                              fontWeight: FontWeight.w800,
-                              fontStyle: FontStyle.italic,
-                              color: Colors.blue)),
+                          // TextStyle specific to this widget
+                          style: Theme.of(context).textTheme.bodyText1),
                     ],
                   ),
                 ),
+                Divider(),
                 Container(
-                  padding: EdgeInsets.all(15),
                   child: Text(issue.shortTitle,
                       style: Theme.of(context).textTheme.headline6),
                 ),
                 Container(
-                  padding: EdgeInsets.all(10),
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.spaceBetween,
                     children: <Widget>[
@@ -90,64 +84,6 @@ class IssueListItem extends StatelessWidget {
             ),
           ),
         ),
-      ),
-    );
-  }
-}
-
-class VotingStatusWidget extends StatelessWidget {
-  Map issuesMap;
-  bool voted;
-  VotingStatusWidget(Map m, bool v) {
-    this.issuesMap = m;
-    this.voted = v;
-  }
-
-  statusMessage() {
-    String s = "Closed";
-    Color c = Colors.red;
-    var i = Icons.adjust;
-    if (voted) {
-      s = "Voted";
-      c = Colors.blue;
-      i = Icons.check_circle_outline;
-    } else {
-      if (issuesMap["Chamber"] == "House") {
-        if (issuesMap["Passed Senate"] == "") {
-          s = "Open";
-          c = Colors.green;
-          i = Icons.add_circle_outline;
-        }
-      } else {
-        if (issuesMap["Passed House"] == "") {
-          s = "Open";
-          c = Colors.green;
-          i = Icons.add_circle_outline;
-        }
-      }
-    }
-
-    return [c, s, i];
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Column(
-        children: <Widget>[
-          Icon(
-            statusMessage()[2],
-            color: statusMessage()[0],
-            size: 20,
-          ),
-          Text(
-            statusMessage()[1],
-            style: TextStyle(
-                fontSize: 8,
-                fontWeight: FontWeight.bold,
-                color: statusMessage()[0]),
-          ),
-        ],
       ),
     );
   }
